chore(batcher): split driver integration tests into focused files
Break the monolithic tests/driver.rs (915 lines) into five files,
one per logical concern, each as its own test binary:

- tests/reorg.rs      — add_block reorg, L2 reorg event (3 tests)
- tests/throttle.rs   — DA throttle: high/zero backlog, redundancy,
                        Step strategy, active→inactive (5 tests)
- tests/heads.rs      — L1 head source, safe head watch, sender-drop
                        busyloop guard (5 tests)
- tests/lifecycle.rs  — source exhaustion, flush, drain timeout (3 tests)
- tests/pause_resume.rs — pause resets pipeline, resume triggers catchup,
                          paused drops Block/Flush events (3 tests)

Update Cargo.toml with one [[test]] entry per file, each gated on
required-features = ["test-utils"].
